alphas should be floats

+++ postprocessing.py
@@ -266,7 +266,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Parser')
     parser.add_argument('--model_prefix', type=str)
     parser.add_argument('--model_idx', type=int, nargs='+')
-    parser.add_argument('--alphas', type=int, nargs='+')
+    parser.add_argument('--alphas', type=float, nargs='+')
     args = parser.parse_args()
 
     # plot_mse_llk(model_name)
